+++ src/app/dashboard/layout.tsx
@@ -3,6 +3,7 @@ import React, { useEffect, useState } from "react";
 import useSession from "@/hooks/useSession";
 import Sidebar from "./components/Sidebar";
 import NavbarDashboard from "./components/NavbarDashboard";
+import NavbarPizarra from "./pizarra/components/NavbarPizarra";
 import Spinner from "@/components/Spinner";
 import { DashboardUIProvider, useDashboardUI } from "./ui";
 import {
@@ -67,6 +68,7 @@ function ProtectedDashboard({ children }: { children: React.ReactNode }) {
   // Altura del navbar
   const navbarHeight = NAVBAR_HEIGHT;
   const isAlmacenDetail = /^\/dashboard\/almacenes\/\d+(\/|$)/.test(pathname);
+  const isPizarra = pathname.startsWith('/dashboard/pizarra');
 
   return (
     <div
@@ -87,7 +89,13 @@ function ProtectedDashboard({ children }: { children: React.ReactNode }) {
           }}
           data-oid="taw.mzt"
         >
-          {usuario && <NavbarDashboard usuario={usuario} data-oid="m6qmdem" />}
+          {usuario && (
+            isPizarra ? (
+              <NavbarPizarra />
+            ) : (
+              <NavbarDashboard usuario={usuario} data-oid="m6qmdem" />
+            )
+          )}
         </div>
       )}
